Upgraded Sharp Architecture to use Fluent NHibernate 1.0 (Rev 593), NHibernate 2.1.0.4000 and NHibernate Linq 1.0 (Rev 4000).  VS template and Northwind sample application were also updated.
Added .gitignore file at project level.

Added an option to the Nant build that copies all updated files to the VS template solution directory and rebuilds the VS project template.  This may have resulted in putting some files into the template zip that are not strictly necessary.  However, this can be refined easily.