Remove StaleResults parameter from IRavenRepo
Removed as of now, as similar functionality can be configured on the
DocumentStore using:
Conventions.DefaultQueryingConsistency=ConsistencyOptions.MonotonicRead;

Using WaitForNonStaleResults is not recommended as that might cause a
long wait/timeouts if there are many writes.

If we are going to expose anyting we should at least expost the whole
customisation, not just a bool. Dont think we need to though.